Stances + Sprint bandaids
Bandaid fix to stances to result in just Shii-Cho being available. See
Issues. Also changed stance animation.
Fixed being able to 'sprint' while in spectator (fov change only).
Bandaid fix to being able to sprint while in the air, eez started a much
more dynamic fix to this but it doesn't appear to be complete and/or
working atm so this will do for now.