Fix #34 - buy button works reliably now.
The problem was when the player went to buy an item. The server would,
each time, attempt to identify the vendor the player was buying from.
However this occasionally gave different results due to the way it was
doing this.

The vendor the player is using is now saved when he/she views the vendor's
items, and will always be correct when trying to buy.

This commit also allows players to view items they do not have enough
credits to buy. Further work is needed to grey out the buy button
(although the server prevents the player from buying items which cost too
much anyway).