NPC Body Que Crash
I love to spam spawn NPCs to test the engines limits, as it stands the
game doesn't remove NPCs, possibly due to the intent of the future
looting system. I added a cvar to enable dead bodys to be removed along
with 2 extra cvars that can set the time of removal, one for droids and
another for everything else. Best leave droids at 0 for the most part.

The cvar as it stands should not break original functionality.

PS - JKG handles large amounts of NPCs surprisingly well but not the
saber system, a large amount of saber NPCs will trigger a break  at
"VectorCopy(end,
self->client->saber[saberNum].blade[rBladeNum].trail.tip);" in wp_saber