Implemented redsaurus Player Species Fix
A nice little fix that is on OpenJK.
In terms of RAM my machine has noticed a decrease of over 60mb (274mb ->
211mb) on the client side. This most likly would scale alot higher if
JKG had more player models. Thanks redsaurus

An extra thing to note, I added a PLAYER_SPEC_MAX define to allow you to
scale up or down, if JKG has missing species in the future, its most
likly because this number is too low, to many models (even ones that are
not in playerchoice.txt) may fail to load into the species lists. I
doubled this value from its orginal amount 2048 -> 4096