Fix demos having an extra 2 bytes in their first frame
Causing an issue where they wouldn't load properly.  You will need new
demos going forward since the extra 2 bytes is gone.  Was introduced in
the record function from the old RMG code as a fallback legacy but it
was gutted entirely in JKG except for this part.

Also fixing offsets using a regular int in the netfields.  size_t must
be used for compatibility with 64-bit.