Fix Travis CI macOS build (#312)
For some reason on the Travis CI macOS builds, `PNG_INCLUDE_DIRS` is empty even when it should be identical to `PNG_INCLUDE_DIR`. As a result compilation fails since `png.h` cannot be found.

This PR adds `PNG_INCLUDE_DIR` to the include directory which does have the include path for `png.h`. Since `PNG_INCLUDE_DIR` is marked for deprecation and may stop working, we continue to use `PNG_INCLUDE_DIRS` as well since this is documented as the correct variable to use for include directories.